+++ src/core/messenger/type.ts
@@ -1,5 +1,5 @@
 import { isString } from '@guardian/libs';
-import fastdom from '../../utils/fastdom-promise';
+import fastdom from '../../lib/fastdom-promise';
 import type { RegisterListener } from '../messenger';
 
 const setType = (adSlotType: string, adSlot: Element) =>