+++ src/init/consented/dynamic-ad-slots.ts
@@ -1,11 +1,10 @@
-import { init as initArticleAsideAdverts } from '../../insert/article-aside-adverts';
 import { initCommentsExpandedAdverts } from '../../insert/comments-expanded-advert';
 import { init as initFootballRightAds } from '../../insert/fixures';
 import { init as initHighMerch } from '../../insert/high-merch';
 import { init as initMobileCrosswordsAdvert } from '../../insert/mobile-crossword-banner';
 import { init as initMobileSticky } from '../../insert/mobile-sticky';
 import { init as initLiveblogAdverts } from '../../insert/spacefinder/liveblog-adverts';
-import { reportError } from '../../utils/report-error';
+import { reportError } from '../../lib/error/report-error';
 import { initArticleBodyAdverts } from './article-body-adverts';
 
 type Modules = Array<[`${string}-${string}`, () => Promise<unknown>]>;
@@ -13,7 +12,6 @@ type Modules = Array<[`${string}-${string}`, () => Promise<unknown>]>;
 const dynamicAdSlotModules: Modules = [
 	['cm-mobileSticky', initMobileSticky],
 	['cm-highMerch', initHighMerch],
-	['cm-articleAsideAdverts', initArticleAsideAdverts],
 	['cm-articleBodyAdverts', initArticleBodyAdverts],
 	['cm-liveblogAdverts', initLiveblogAdverts],
 	['cm-commentsExpandedAdverts', initCommentsExpandedAdverts],