+++ src/init/consented/static-ad-slots.ts
@@ -10,7 +10,7 @@ import { commercialFeatures } from '../../lib/commercial-features';
 import { getCurrentBreakpoint } from '../../lib/detect/detect-breakpoint';
 import { dfpEnv } from '../../lib/dfp/dfp-env';
 import { queueAdvert } from '../../lib/dfp/queue-advert';
-import { isInUk, isInUsa } from '../../utils/geo-utils';
+import { isInUk, isInUsa } from '../../lib/geo/geo-utils';
 import { setupPrebidOnce } from './prepare-prebid';
 import { removeDisabledSlots } from './remove-slots';
 
@@ -54,7 +54,6 @@ const decideAdditionalSizes = (adSlot: HTMLElement): SizeMapping => {
  * Static ad slots that were rendered on the page by the server are collected here.
  *
  * For dynamic ad slots that are created at runtime, see:
- *  - article-aside-adverts
  *  - article-body-adverts
  *  - high-merch
  */