+++ src/insert/high-merch.ts
@@ -1,11 +1,13 @@
 import { createAdSlot, wrapSlotInContainer } from '../core/create-ad-slot';
 import { commercialFeatures } from '../lib/commercial-features';
-import fastdom from '../utils/fastdom-promise';
+import fastdom from '../lib/fastdom-promise';
 
 /**
  * Initialise merchandising-high ad slot on Frontend rendered content
  *
  * On DCR, these ad slots are server side rendered
+ *
+ * Revisit whether this code is needed once galleries have been migrated to DCR
  */
 export const init = (): Promise<void> => {
 	if (commercialFeatures.highMerch) {