+++ src/insert/spacefinder/carrot-traffic-driver.ts
@@ -1,7 +1,7 @@
 import { createAdSlot } from '../../core/create-ad-slot';
 import { commercialFeatures } from '../../lib/commercial-features';
 import { getCurrentTweakpoint } from '../../lib/detect/detect-breakpoint';
-import fastdom from '../../utils/fastdom-promise';
+import fastdom from '../../lib/fastdom-promise';
 import { fillDynamicAdSlot } from '../fill-dynamic-advert-slot';
 import { spaceFiller } from './space-filler';
 import type { SpacefinderRules, SpacefinderWriter } from './spacefinder';