+++ src/lib/dfp/non-refreshable-line-items.spec.ts
@@ -1,10 +1,10 @@
-import { reportError } from '../../utils/report-error';
+import { reportError } from '../../lib/error/report-error';
 import {
 	fetchNonRefreshableLineItemIds,
 	memoizedFetchNonRefreshableLineItemIds,
 } from './non-refreshable-line-items';
 
-jest.mock('../../utils/report-error', () => ({
+jest.mock('../../lib/error/report-error', () => ({
 	reportError: jest.fn(),
 }));