+++ src/lib/header-bidding/prebid/appnexus.spec.ts
@@ -4,7 +4,7 @@ import {
 	isInRow as isInRow_,
 	isInUk as isInUk_,
 	isInUsa as isInUsa_,
-} from '../../../utils/geo-utils';
+} from '../../../lib/geo/geo-utils';
 import {
 	containsBillboard as containsBillboard_,
 	containsBillboardNotLeaderboard as containsBillboardNotLeaderboard_,
@@ -31,7 +31,7 @@ const containsMpu = containsMpu_ as jest.Mock;
 const containsMpuOrDmpu = containsMpuOrDmpu_ as jest.Mock;
 const getBreakpointKey = getBreakpointKey_ as jest.Mock;
 
-jest.mock('utils/geo-utils');
+jest.mock('lib/geo/geo-utils');
 const isInAuOrNz = isInAuOrNz_ as jest.Mock;
 const isInRow = isInRow_ as jest.Mock;
 const isInUk = isInUk_ as jest.Mock;