+++ src/lib/header-bidding/prebid/bid-config.spec.ts
@@ -7,7 +7,7 @@ import {
 	isInUk as isInUk_,
 	isInUsa as isInUsa_,
 	isInUsOrCa as isInUsOrCa_,
-} from '../../../utils/geo-utils';
+} from '../../../lib/geo/geo-utils';
 import type { HeaderBiddingSize, PrebidBidder } from '../prebid-types';
 import {
 	containsBillboard as containsBillboard_,
@@ -46,7 +46,7 @@ const {
 	getOzonePlacementId,
 } = _;
 
-jest.mock('lib/build-page-targeting', () => ({
+jest.mock('lib/page-targeting', () => ({
 	buildAppNexusTargeting: () => 'someTestAppNexusTargeting',
 	buildAppNexusTargetingObject: () => 'someAppNexusTargetingObject',
 	getPageTargeting: () => 'bla',
@@ -72,7 +72,7 @@ const stripMobileSuffix = stripMobileSuffix_ as jest.Mock;
 const getBreakpointKey = getBreakpointKey_ as jest.Mock;
 const isUserInVariant = isUserInVariant_ as jest.Mock;
 
-jest.mock('utils/geo-utils');
+jest.mock('lib/geo/geo-utils');
 const isInAuOrNz = isInAuOrNz_ as jest.Mock;
 const isInRow = isInRow_ as jest.Mock;
 const isInUk = isInUk_ as jest.Mock;